Create `network_channels` module
Move `ReplicationChannel` and `NetworkChannels` into it.
