+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Rename `ClientCache` into `ConnectedClients`.
 - Rename `replicon_core` module into `core`.
+- Move `ReplicationChannel` and `NetworkChannels` into `network_channels` module.
 
 ## [0.23.0] - 2024-02-22
